properly refreshes pages
