+++ clientui/src/components/Modals/TrialModal.jsx
@@ -1,13 +1,13 @@
-
+import { useState } from 'react';
+import Modal from 'react-bootstrap/Modal';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
 import Card from 'react-bootstrap/Card';
 import ListGroup from 'react-bootstrap/ListGroup';
-import { Link } from 'react-router-dom';
 import InputGroup from 'react-bootstrap/InputGroup';
-import { useState } from 'react';
+import { Link } from 'react-router-dom';
 
-export default function TrialModal({setId, closeModal}){
+export default function TrialModal({ show, closeModal, setId }) {
     const [maxTrials, setMaxTrials] = useState(0);
     const [numberOfCards, setNumberOfCards] = useState(0);
 
@@ -20,39 +20,39 @@ export default function TrialModal({setId, closeModal}){
     };
 
     return (
-        <div>
-             <Card style={{ width: '18rem' }}>
-                
-                <ListGroup className="list-group-flush">
-                    <InputGroup className="mb-3">
-                        <InputGroup.Text >Trials:</InputGroup.Text>
+        <Modal show={show} onHide={closeModal} centered>
+            <Modal.Header closeButton>
+                <Modal.Title>Trial Configuration</Modal.Title>
+            </Modal.Header>
+            <Modal.Body>
+                <Card style={{ width: '18rem' }}>
+                    <ListGroup className="list-group-flush">
+                        <InputGroup className="mb-3">
+                            <InputGroup.Text>Trials:</InputGroup.Text>
                             <Form.Control
-                            type='number'
-                            value={maxTrials}
-                            onChange={handleMaxTrialsChange}
+                                type='number'
+                                value={maxTrials}
+                                onChange={handleMaxTrialsChange}
                             />
-                    </InputGroup>
+                        </InputGroup>
 
-                    <InputGroup className="mb-3">
-                        <InputGroup.Text >Cards on screen:</InputGroup.Text>
+                        <InputGroup className="mb-3">
+                            <InputGroup.Text>Cards on screen:</InputGroup.Text>
                             <Form.Control
-                            type='number'
-                            value={numberOfCards}
-                            onChange={handleNumberOfCardsChange}
+                                type='number'
+                                value={numberOfCards}
+                                onChange={handleNumberOfCardsChange}
                             />
-                    </InputGroup>
-                </ListGroup>
-                <Card.Body>
-                    <Link to={`/${setId}/trial?maxTrials=${maxTrials}&numberOfCards=${numberOfCards}`}>
-                        <Button variant="primary">
-                            Begin
-                        </Button>{' '}
-                    </Link>
-                    <Button onClick={closeModal}>
-                        Exit
-                    </Button>
-                </Card.Body>
-            </Card>
-        </div>
-    )
+                        </InputGroup>
+                    </ListGroup>
+                </Card>
+            </Modal.Body>
+            <Modal.Footer>
+                <Link to={`/${setId}/trial?maxTrials=${maxTrials}&numberOfCards=${numberOfCards}`}>
+                    <Button variant="primary">Begin</Button>
+                </Link>
+                <Button variant="secondary" onClick={closeModal}>Exit</Button>
+            </Modal.Footer>
+        </Modal>
+    );
 }
